--- a/src/truthy.ml
+++ b/src/truthy.ml
@@ -1,0 +1,5 @@
+type 'v t = 'v
+
+module Unsafe = struct
+  let create x = x
+end