--- a/src/truthy.mli
+++ b/src/truthy.mli
@@ -1,0 +1,5 @@
+type 'v t = private 'v
+
+module Unsafe : sig
+  val create : 'v -> 'v t
+end