This PR
* adds `Truthy` values type (0 - false, anything else - true) and the natural cheap operations that may be done with them
* adds a zero-constraint `not_equal` function for field elements
* adds a zero-constraint `not_equal` function to `Bitstring_checked`
* adds `Bitstring_checked.Assert.either_equal`
  - this costs the same as `Bitstring_checked.Assert.either`, but it lets us make conditional assertions
  - this also doesn't go through `Boolean.t`, so we don't have to compute the inverses of any field elements

The biggest motivator for this is the use of `Schnorr.Checked.verifies` in Coda, where
* we use 2 constraints for each chunk to create its `Boolean.t`
* we calculate the field inverse for each chunk to satisfy the constraints for each `Boolean.t`
* we use a further 2 constraints for the return value
* we use a further constraint and compute another field inverse for the final assertion in `Transaction_snark.check_signature`

By contrast, with this we could stop computing any inverses (which showed this up as a surprising hotspot during profiling) and more than halve the number of constraints for the assertion along the way.

Concretely, the changed code might look like:
```
    let%snarkydef assert_either_verifies ~or_ shifted ((_, h) as signature) pk m =
      verification_hash shifted signature pk m
      >>= Curve.Scalar.Checked.Assert.either_equal ~or_ h
```
```
    module Assert = struct
      ...
      let either_equal : or_:Boolean.t -> var -> var -> (unit, _) Checked.t =
       fun ~or a b ->
        Bitstring_checked.Assert.either_equal
          ~or_not:(Truthy.of_boolean (Boolean.not or_))
          (Bitstring.Lsb_first.to_list a)
          (Bitstring.Lsb_first.to_list b)
    end
```