@imeckler 2019-07-15
Cool idea! Surely the inversions performed in curve operations are where the bulk of the inversions are happening though no?