@mrmr1993 2019-07-15
> Surely the inversions performed in curve operations are where the bulk of the inversions are happening though no?

This is true, but the `equal` call was turning out much heavier than I had expected, and it seemed like something we could fix. It could also be that there's some inlining happening there that's confusing the profiler.

I'll run some tests to see if we see much difference.